Fix error from opening attachments in tmux
Problem:

Mac OS X's built-in `open` command is used to open email attachments
in the appropriate application.

`open` has some incompatibilities with tmux -
in the case of our Mutt setup, it raises the error:

```
The window server could not be contacted.
open must be run with a user logged in at the console, either as that user or as root.
```

Solution:

This is the kind of problem that `reattach-to-user-namespace` is built
to solve. (https://github.com/ChrisJohnsen/tmux-MacOSX-pasteboard)

Running the `open` commands through `reattach-to-user-namespace`
resolves the issue.