Change git comment character to support gitmessage
## Problem:

We could not embed Markdown headers in commit messages
because git treats lines starting with the `#` character as a comment.

## Solution:

Add a git config entry to change the character used for comments.

This will break vim's syntax highlighting for git commit messages.
To fix it, edit `$VIMRUNTIME/syntax/gitcommit.vim`,
and update the `syn match gitCommitComment` entry to `"^|.*"`.